requirements.txt
GitPython
openai
python-dotenv
